Add files via upload
Coding is done for the most part and only some revising is needed in case something goes wrong with a new text set --> Original codes are included in the "development" file
I am currently working on the draft: adding big and small sections and presenting codes and explaining methods etc. I will update this to the latest version before we meet.